Added Sky Manager

+++ Trex/Entities/SkyManager.cs
@@ -7,10 +7,11 @@ using System.Text;
 
 namespace TrexRunner.Entities
 {
-    public class SkyManager : IGameEntity
+    public class SkyManager : IGameEntity, IDayNightCycle
     {
         private const int CLOUD_DRAW_ORDER = -1;
-        private const int STAR_DRAW_ORDER = -2;
+        private const int STAR_DRAW_ORDER = -3;
+        private const int MOON_DRAW_ORDER = -2;
 
 
         private const int CLOUD_MIN_POS_Y = 20;
@@ -18,6 +19,8 @@ namespace TrexRunner.Entities
 
         private const int STAR_MIN_POS_Y = 10;
         private const int STAR_MAX_POS_Y = 60;
+
+        private const int MOON_POS_Y = 30;
 
         private const int CLOUD_MIN_DISTANCE = 100;
         private const int CLOUD_MAX_DISTANCE = 400;
@@ -29,6 +32,7 @@ namespace TrexRunner.Entities
         private readonly ScoreBoard _scoreBoard;
         private readonly Trex _trex;
         private Texture2D _spriteSheet;
+        private Moon _moon;
 
         private int _targetCloudDistance;
         private int _targetStarDistance;
@@ -37,6 +41,9 @@ namespace TrexRunner.Entities
 
         public int DrawOrder => 0;
 
+        public int NightCount { get; private set; }
+
+        public bool isNight { get; }
 
         public SkyManager(Trex trex, Texture2D spriteSheet, EntityManager entityManager, ScoreBoard scoreBoard)
         {
@@ -47,6 +54,8 @@ namespace TrexRunner.Entities
             _random = new Random();
         }
 
+
+
         public void Draw(SpriteBatch spriteBatch, GameTime gameTime)
         {
             
@@ -54,12 +63,26 @@ namespace TrexRunner.Entities
 
         public void Update(GameTime gameTime)
         {
+            if (_moon == null)
+            {
+                _moon = new Moon(this, _spriteSheet, _trex, new Vector2(TRexGame.WINDOW_WIDTH, MOON_POS_Y));
+                _moon.DrawOrder = MOON_DRAW_ORDER;
+                _entityManager.AddEntity(_moon);
+            }
+
             HandleCloudSpawning();
             HandleStarSpawning();
 
             foreach(SkyObject skyObject in _entityManager.GetEntitiesOfType<SkyObject>().Where(s => s.Position.X < -200))
             {
-                _entityManager.RemoveEntity(skyObject);
+                if(skyObject is Moon moon)
+                {
+                    moon.Position = new Vector2(TRexGame.WINDOW_WIDTH, MOON_POS_Y);
+                }
+                else
+                {
+                    _entityManager.RemoveEntity(skyObject);
+                }
             }
         }
